+++ pkg/plugin/sdk/v1alpha1/plugin/interface.go
@@ -18,8 +18,4 @@ type Plugin interface {
 		isInteractiveMode bool,
 		promptRunner ioutils.PromptRunner,
 	) error
-	PostTestHook(
-		isInteractiveMode bool,
-		promptRunner ioutils.PromptRunner,
-	) error
 }
