+++ pkg/plugin/sdk/v1alpha2/plugin/grpc.go
@@ -74,6 +74,25 @@ func (m *GRPCServer) PostBuildHook(
 		m.Impl.PostBuildHook(req.IsInteractiveMode, prompter)
 }
 
+// PostTestHook translates the gRPC call to the Plugin PostTestHook
+// implementation. It starts a prompt runner that is passed to the Plugin
+// instance to be able to perform prompt actions to the CLI user.
+func (m *GRPCServer) PostTestHook(
+	ctx context.Context,
+	req *proto.PostTestHookReq,
+) (*proto.PostTestHookRes, error) {
+	conn, err := m.broker.Dial(req.BrokerId)
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+
+	client := proto.NewPrompterClient(conn)
+	prompter := &PrompterGRPCClient{client: client}
+	return &proto.PostTestHookRes{},
+		m.Impl.PostTestHook(req.IsInteractiveMode, prompter)
+}
+
 // GRPCClient implements the gRPC client that is used by the Core to communicate
 // with the Plugin instances.
 type GRPCClient struct {
@@ -105,6 +124,27 @@ func (m *GRPCClient) PostBuildHook(isInteractiveMode bool, promptRunner ioutils.
 		IsInteractiveMode: isInteractiveMode,
 	}
 	_, err := m.client.PostBuildHook(context.Background(), req)
+	s.Stop()
+	return err
+}
+
+// PostTestHook is called from the Core to execute the Plugin PostTestHook. It
+// starts the prompt runner server with the provided PromptRunner.
+func (m *GRPCClient) PostTestHook(isInteractiveMode bool, promptRunner ioutils.PromptRunner) error {
+	prompterServer := &PrompterGRPCServer{promptRunner: promptRunner}
+	var s *grpc.Server
+	serverFunc := func(opts []grpc.ServerOption) *grpc.Server {
+		s = grpc.NewServer(opts...)
+		proto.RegisterPrompterServer(s, prompterServer)
+		return s
+	}
+	brokerID := m.broker.NextId()
+	go m.broker.AcceptAndServe(brokerID, serverFunc)
+	req := &proto.PostTestHookReq{
+		BrokerId:          brokerID,
+		IsInteractiveMode: isInteractiveMode,
+	}
+	_, err := m.client.PostTestHook(context.Background(), req)
 	s.Stop()
 	return err
 }
