+++ pkg/plugin/sdk/v1alpha2/plugin/interface.go
@@ -18,4 +18,8 @@ type Plugin interface {
 		isInteractiveMode bool,
 		promptRunner ioutils.PromptRunner,
 	) error
+	PostTestHook(
+		isInteractiveMode bool,
+		promptRunner ioutils.PromptRunner,
+	) error
 }
