+++ pkg/plugin/system/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
         "//pkg/aspecterrors",
         "//pkg/interceptors",
         "//pkg/ioutils",
-        "//pkg/plugin/sdk/v1alpha1/config",
-        "//pkg/plugin/sdk/v1alpha1/plugin",
+        "//pkg/plugin/sdk/v1alpha2/config",
+        "//pkg/plugin/sdk/v1alpha2/plugin",
         "//pkg/plugin/system/bep",
         "@com_github_hashicorp_go_hclog//:go-hclog",
         "@com_github_hashicorp_go_plugin//:go-plugin",
